Restructured project, support different loss functions and single/multi-label classification within config, same model can be used for all four cases that exist when combining (sl vs ml) and (logits vs. final_activation)

+++ model/loss.py
@@ -23,7 +23,7 @@ def BCE(output, target):
 # This contains Sigmoid itself
 def BCE_with_logits(output, target):
     loss = BCEWithLogitsLoss()
-    return loss(output, target)
+    return loss(output, target.float())
 
 
 # This contains Sigmoid itself
